% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_control.R
\name{cusum_control}
\alias{cusum_control}
\title{cusum_control}
\usage{
cusum_control(
  x,
  target = NULL,
  std_dev = NULL,
  desired_shift = 1,
  k = 0.5,
  h = 5
)
}
\arguments{
\item{x}{input vector}

\item{target}{target value for comparison, the mean of x will be used if missing}

\item{std_dev}{target value for standard deviation, the screened moving range of x will be used if missing}

\item{desired_shift}{how many standard deviations do you want to detect, between 0.5 and 1 is the usual default}

\item{k}{allowable slack - defaults to half of standard deviation}

\item{h}{control limits usually between 4 and 5}
}
\value{
dataframe showing original inputs and calculated control limits
}
\description{
cusum_control
}
\examples{
test_vec3 <- c(1,1,2,3,5,7,11,7,5,7,8,9,5)
controls <- cusum_control(test_vec3, target = 4)


}
