% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_simple_control.R
\name{cusum_control_plot}
\alias{cusum_control_plot}
\title{cusum_control_plot}
\usage{
cusum_control_plot(df, title_text = NULL)
}
\arguments{
\item{df}{input data frame generated by cusum_control function}

\item{title_text}{optional title for chart}
}
\value{
ggplot2 object suited for further customisation
}
\description{
cusum_control_plot
}
\examples{
test_vec3 <- c(1,1,2,3,5,7,11,7,5,7,8,9,5)
controls <- cusum_control(test_vec3, target = 4)
cusum_control_plot(controls)

}
