% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_simple.R
\name{cusum_simple}
\alias{cusum_simple}
\title{cusum_simple}
\usage{
cusum_simple(x, target = NULL)
}
\arguments{
\item{x}{a numeric vector from which to calculate the cumulative sum statistics}

\item{target}{value to compare each element of x to. If not provided, the mean of
x will be calculated and used as a target value}
}
\value{
a vector of the cumulative sum statistic, centred on the target value
}
\description{
cusum_simple
}
\examples{

test_vec <- c(0.175, 0.152, 0.15, 0.207, 0.136, 0.212, 0.166)
cusum_simple(test_vec)

}
