% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_single_median_df.R
\name{cusum_single_median_df}
\alias{cusum_single_median_df}
\title{cusum_single_median_df}
\usage{
cusum_single_median_df(x, target = NULL)
}
\arguments{
\item{x}{a numeric vector from which to calculate the cumulative sum statistics}

\item{target}{value to compare each element of x to. If not provided, the median value of
x will be calculated and used as a target value}
}
\value{
a dataframe with the original values, target, the variance, the cumulative sum of the variance, and the cumulative sum centered on the target value. This centering is achieved by adding the target value to the cumulative sum.
}
\description{
cusum_single_median_df
}
\examples{

test_vec <- c(0.175, 0.152, 0.15, 0.207, 0.136, 0.212, 0.166)
cusum_single_median_df(test_vec, target = 0.16)
cusum_single_median_df(test_vec)


}
