% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cusumcharter-deprecated}
\alias{cusumcharter-deprecated}
\title{Deprecated functions in cusumcharter}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link{cusum_single_median}}: This function is deprecated, and will
be removed in the next version of this package.
}
}
